calculate score by gene

+++ snpScore.py
@@ -2,6 +2,7 @@ import fire
 import gzip
 import sys
 import delegator
+import gtfparse
 import pandas as pd
 import numpy as np
 from pathlib import Path
@@ -141,12 +142,12 @@ def slidewindow(obj, window, step):
         yield obj[i:i+window]
 
 
-def make_snp_number_windows(stat_df, window, step, outdir, force):
-    snp_num_window_file = outdir / f'snp_num.window.w{window}.s{step}.bed'
+def make_snp_number_windows(stat_df, stat_label, window, step, outdir, force):
+    snp_num_window_file = outdir / f'snp_num.{stat_label}.window.w{window}.s{step}.bed'
     if (not snp_num_window_file.exists()) or force:
         logger.info(
-            'making snp number slidewindow bed file windows {w} step {s}',
-            w=window, s=step)
+            'making {l} snp number slidewindow bed file windows {w} step {s}...',
+            w=window, s=step, l=stat_label)
         snp_num_window_list = []
         for slidewindow_i in slidewindow(stat_df.index, window, step):
             chrom = stat_df.Chrom[slidewindow_i].unique()
@@ -159,6 +160,7 @@ def make_snp_number_windows(stat_df, window, step, outdir, force):
             snp_num_window_list.append([score_chrom, start, end])
         snp_num_window_df = pd.DataFrame(snp_num_window_list,
                                          columns=['Chrom', 'Start', 'End'])
+        snp_num_window_df.loc[:, 'label'] = snp_num_window_df.index.map(lambda x: f'snp_group{x+1}')
         snp_num_window_df.to_csv(
             snp_num_window_file, sep='\t', index=False, header=False)
     return snp_num_window_file
@@ -168,30 +170,72 @@ def make_genome_windows(chr_file, window, step, outdir, force):
     window_file = outdir / f'genome.window.w{window}.s{step}.bed'
     if (not window_file.exists()) or force:
         logger.info(
-            'making genome slidewindow bed file windows {w} step {s}',
+            'making genome slidewindow bed file windows {w} step {s}...',
             w=window, s=step)
         cmd = f'bedtools makewindows -g {chr_file} -w {window} -s {step} > {window_file}'
         delegator.run(cmd)
+        window_df = pd.read_csv(window_file, sep='\t', header=None,
+                                names=['Chrom', 'Start', 'End'])
+        window_df.loc[:, 'Label'] = window_df.apply(
+            lambda x: '{0}:{1}-{2}'.format(
+                x['Chrom'], x['Start'], x['End']), axis=1)
+        window_df.to_csv(window_file, sep='\t', index=False,
+                         header=False)
+    return window_file
+
+
+def make_gene_window(gtf, outdir, force, by='gene'):
+    window_file = outdir / f'{by}.window.bed'
+    if (not window_file.exists()) or force:
+        logger.info(
+            'making {by} bed file...',
+            by=by)
+        gtf_df = gtfparse.read_gtf(gtf)
+        if by == 'gene':
+            if 'gene' in gtf_df.feature.unique():
+                pass
+            else:
+                if 'exon' not in gtf_df.feature.unique():
+                    sys.exit('exon not in gtf file!')
+                exon_gtf = gtf_df[gtf_df.feature == 'exon']
+                gene_start = exon_gtf.groupby(['gene_id'])['start'].min() - 1
+                gene_end = exon_gtf.groupby(['gene_id'])['end'].max()
+                gene_chrom = exon_gtf.loc[
+                    :, ['gene_id', 'seqname']
+                ].drop_duplicates().set_index('gene_id')
+                gene_df = pd.concat(
+                    [gene_chrom, gene_start, gene_end],
+                    axis=1, sort=False)
+                gene_df = gene_df.reset_index()
+                gene_df.sort_values(['seqname', 'start', 'end'], inplace=True)
+                gene_df.to_csv(window_file, sep='\t',
+                               index=False, header=False,
+                               columns=['seqname', 'start', 'end', 'index'])
     return window_file
 
 
 def varscore_by_window(stat_df,
                        window_file,
                        outdir,
+                       stat_label,
+                       by,
                        est=None, method='var',
                        offset=0.00001):
+    logger.info('calculating score based on {stat_label} snp by {by} using {method}...',
+                by=by, method=method, stat_label=stat_label)
     groups = stat_df.columns[3:-1]
-    alt_freq_stat_bed = outdir / 'snp.plot.bed'
-    alt_freq_stat_df = stat_df.copy()
-    alt_freq_stat_df.loc[:, 'start'] = alt_freq_stat_df.Pos - 1
-    bed_cols = ['Chrom', 'start', 'Pos']
-    bed_cols.extend(groups)
-    alt_freq_stat_df.to_csv(alt_freq_stat_bed, sep='\t',
-                            columns=bed_cols, header=None, index=False)
+    alt_freq_stat_bed = outdir / f'snp.{stat_label}.plot.bed'
+    if not alt_freq_stat_bed.exists():
+        alt_freq_stat_df = stat_df.copy()
+        alt_freq_stat_df.loc[:, 'start'] = alt_freq_stat_df.Pos - 1
+        bed_cols = ['Chrom', 'start', 'Pos']
+        bed_cols.extend(groups)
+        alt_freq_stat_df.to_csv(alt_freq_stat_bed, sep='\t',
+                                columns=bed_cols, header=None, index=False)
     window_bed = BedTool(str(window_file))
     snp_bed = BedTool(str(alt_freq_stat_bed))
     intersect_obj = window_bed.intersect(snp_bed, sorted=True, wo=True)
-    intersect_obj_cols = ['Chrom', 'Start', 'End']
+    intersect_obj_cols = ['Chrom', 'Start', 'End', 'Label']
     intersect_obj_cols.extend(['snp_Chrom', 'snp_start', 'snp_end'])
     intersect_obj_cols.extend(groups)
     intersect_obj_cols.append('overlap')
@@ -201,11 +245,11 @@ def varscore_by_window(stat_df,
     intersect_df.drop(['snp_Chrom', 'snp_start', 'snp_end',
                        'overlap'], axis=1, inplace=True)
     varscore_size_df = intersect_df.groupby(
-        ['Chrom', 'Start', 'End']).size()
+        ['Chrom', 'Start', 'End', 'Label']).size()
     if method == 'var':
         mask = varscore_size_df > 1
         varscore_df = intersect_df.groupby(
-            ['Chrom', 'Start', 'End']).agg(
+            ['Chrom', 'Start', 'End', 'Label']).agg(
                 lambda x: np.var(x) + offset)
         varscore_df = varscore_df[mask]
     elif method == 'est':
@@ -213,13 +257,13 @@ def varscore_by_window(stat_df,
             intersect_df.loc[:, group_i] = intersect_df.loc[
                 :, group_i] - est[n]
         varscore_df = intersect_df.groupby(
-            ['Chrom', 'Start', 'End']).agg(
+            ['Chrom', 'Start', 'End', 'Label']).agg(
                 lambda x: np.average(np.power(x, 2))
         )
     elif method == 'snp_index':
         mask = varscore_size_df >= 10
         varscore_df = intersect_df.groupby(
-            ['Chrom', 'Start', 'End']).agg('mean')
+            ['Chrom', 'Start', 'End', 'Label']).agg('mean')
         varscore_df = varscore_df[mask]
     else:
         sys.exit('Wrong analysis method.')
@@ -243,7 +287,7 @@ def score_plot(score_file, method, force):
         cmd = f'Rscript {SNP_SCORE_PLOT} --input {score_file} --output {out_prefix}'
     elif method == 'snp_index':
         out_plot = score_file.with_suffix('')
-        cmd = f'Rscript {SNP_INDEX_PLOT} --snp_index_file {score_file} --plot_dir {out_plot} --name {plot_name}'
+        cmd = f'Rscript {SNP_INDEX_PLOT} --snp_index_file {score_file} --plot_dir {out_plot}'
     else:
         sys.exit(f'Wrong snp score method [{method}]!')
     if not out_plot.exists() or force:
@@ -407,33 +451,49 @@ def filter_snp(alt_freq_stat_df, freq_order, outdir, force):
 def cal_varscore(alt_freq_stat_df, freq_est,
                  outdir, force,
                  by='snp_num', method='var',
-                 snp_number_window=5,
-                 snp_number_step=3,
+                 snp_number_window=10,
+                 snp_number_step=5,
                  genome_window=2000000,
                  genome_step=1000000,
                  stat_label='all',
-                 chr_size=None):
+                 chr_size=None,
+                 gtf=None):
     varscore_file = outdir / f'snp.{stat_label}.{by}.{method}.score.txt'
     if (not varscore_file.exists()) or force:
         alt_freq_stat_pos_df = alt_freq_stat_df.iloc[:, 0:3]
         analysis_groups = alt_freq_stat_df.columns[3:]
-        logger.info('calculating score based on {stat_label} snp by {by} using {method}...',
-                    by=by, method=method, stat_label=stat_label)
         if by in 'snp_num':
             snp_number_window_file = make_snp_number_windows(
-                alt_freq_stat_df, snp_number_window,
+                alt_freq_stat_df, stat_label, snp_number_window,
                 snp_number_step, outdir, force)
             varscore_df = varscore_by_window(alt_freq_stat_df,
                                              snp_number_window_file,
                                              outdir,
+                                             stat_label,
+                                             by,
                                              est=freq_est,
                                              method=method)
         elif by == 'genome_window':
+            if chr_size is None:
+                return 0
             window_file = make_genome_windows(chr_size, genome_window,
                                               genome_step, outdir, force)
             varscore_df = varscore_by_window(alt_freq_stat_df,
                                              window_file,
                                              outdir,
+                                             stat_label,
+                                             by,
+                                             est=freq_est,
+                                             method=method)
+        elif by == 'gene':
+            if gtf is None:
+                return 0
+            window_file = make_gene_window(gtf, outdir, force)
+            varscore_df = varscore_by_window(alt_freq_stat_df,
+                                             window_file,
+                                             outdir,
+                                             stat_label,
+                                             by,
                                              est=freq_est,
                                              method=method)
         if method in ['var', 'est']:
@@ -459,7 +519,7 @@ def varscore(vcf, outdir, mutant_bulk, mutant_bulk_freq,
              wild_parent=None, wild_parent_freq=None,
              background=None, plot_min_depth=5,
              force=False, f2_freq=0.3, f1_freq=0.1,
-             bg_freq=0.2):
+             bg_freq=0.2, gtf=None):
     # init log
     outdir = Path(outdir)
     outdir.mkdir(parents=True, exist_ok=True)
@@ -482,7 +542,7 @@ def varscore(vcf, outdir, mutant_bulk, mutant_bulk_freq,
     # varscore
     freq_est = [mutant_bulk_freq, wild_bulk_freq,
                 mutant_parent_freq, wild_parent_freq]
-    by = ['snp_num', 'genome_window']
+    by = ['snp_num', 'genome_window', 'gene']
     method = ['var', 'est', 'snp_index']
     freq_df_dict = {'all': alt_freq_stat_df,
                     'filter': alt_freq_stat_filter_df}
@@ -494,7 +554,7 @@ def varscore(vcf, outdir, mutant_bulk, mutant_bulk_freq,
                                           outdir, force, by=by_i,
                                           method=method_i,
                                           stat_label=stat_label,
-                                          chr_size=chr_size)
+                                          chr_size=chr_size, gtf=gtf)
 
 
 if __name__ == '__main__':
